Move the support package URL method to a common location, and add tests.
