+++ src/briefcase/platforms/linux/flatpak.py
@@ -95,17 +95,6 @@ class LinuxFlatpakMixin(LinuxMixin):
 class LinuxFlatpakCreateCommand(LinuxFlatpakMixin, CreateCommand):
     description = "Create and populate a Linux Flatpak."
 
-    def support_package_url(self, support_revision):
-        """The URL of the support package to use for apps of this type.
-
-        AppImage uses Standalone Python builds.
-        """
-        version, datestamp = support_revision.split("+")
-        return (
-            "https://github.com/indygreg/python-build-standalone/releases/download/"
-            f"{datestamp}/cpython-{support_revision}-x86_64-unknown-linux-gnu-install_only.tar.gz"
-        )
-
     def output_format_template_context(self, app: AppConfig):
         """Add flatpak runtime/SDK details to the app template."""
         return {
