+++ tests/platforms/linux/test_LinuxMixin__vendor_details.py
@@ -1,18 +1,8 @@
-from unittest.mock import MagicMock
-
 import pytest
 
-from briefcase.platforms.linux import LinuxMixin, parse_freedesktop_os_release
+from briefcase.platforms.linux import parse_freedesktop_os_release
 
 from .os_release import OS_RELEASE
-
-
-@pytest.fixture
-def linux_mixin():
-    "A Linux mixin with a mocked tools collection"
-    linux_mixin = LinuxMixin()
-    linux_mixin.tools = MagicMock()
-    return linux_mixin
 
 
 @pytest.mark.parametrize(
